Backend MVP complete, gql endpoint complete
