+++ backend/ArrivalTimes.js
@@ -7,54 +7,6 @@
 
 // export const
 
-const gqlArrivals = {
-	value: 0,
-	stops: [
-		{
-			name: '1',
-			buses: {1: [0, 15], 2: [2, 17], 3: [4, 19]}
-		},
-		{
-			name: '2',
-			buses: {1: [2, 17], 2: [4, 19], 3: [6, 21]}
-		},
-		{
-			name: '3',
-			buses: {1: [4, 19], 2: [6, 21], 3: [8, 23]}
-		},
-		{
-			name: '4',
-			buses: {1: [6, 21], 2: [8, 23], 3: [10, 25]}
-		},
-		{
-			name: '5',
-			buses: {1: [8, 23], 2: [10, 25], 3: [12, 27]}
-		},
-		{
-			name: '6',
-			buses: {1: [10, 25], 2: [12, 27], 3: [14, 29]}
-		},
-		{
-			name: '7',
-			buses: {1: [12, 27], 2: [14, 29], 3: [1, 16]}
-		},
-		{
-			name: '8',
-			buses: {1: [14, 29], 2: [1, 16], 3: [3, 18]}
-		},
-		{
-			name: '9',
-			buses: {1: [1, 16], 2: [3, 18], 3: [5, 20]}
-		},
-		{
-			name: '10',
-			buses: {1: [3, 18], 2: [5, 20], 3: [7, 22]}
-		}
-	]
-};
-
-module.exports = {gqlArrivals};
-
 const nextArrival = {
 	0: {
 		1: {1: [0, 15], 2: [2, 17], 3: [4, 19]},
@@ -237,3 +189,5 @@ const nextArrival = {
 		10: {1: [4, 19], 2: [6, 21], 3: [8, 23]}
 	}
 };
+
+module.exports = {nextArrival};
